third_party/blink/web_tests/accessibility/aria-hidden-hides-all-elements.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
</head>
<body id="body">

<div id="content" role="group">

<main id="main" aria-hidden="true">
    <video controls="controls"></video>
    <select id="select"><option>a</select>
    <input type="range">
    <input type="data">
    <progress></progress>
    <ul id="ul"><li>item</li></ul>
    <select size="10"><option>a</select>
    <img src="resources/cake.png" width="145" height="126" usemap="#map">
    <map name="map">
        <area shape="rect" coords="0,0,100,100" href="#" alt="cake">
    </map>
</main>

</div>

<script>
function testChildren(parent, expected_unignored_child_indices) {
    for (let i = 0; i < parent.childrenCount; i++) {
      if (expected_unignored_child_indices.includes(i))
        assert_false(parent.childAtIndex(i).isIgnored);
      else
        assert_true(parent.childAtIndex(i).isIgnored);
    }
}

test(function(t) {
    var content = accessibilityController.accessibleElementById("main");
    assert_equals(content.childrenCount, 17);
    assert_true(content.isIgnored);
    testChildren(content, [0, 2, 4, 6, 11]);
    document.getElementById("ul").tabIndex = -1;

    assert_equals(content.childrenCount, 17,
                      "Making list focusable should not impact if it's hidden");
    assert_true(content.isIgnored);
    testChildren(content, [0, 2, 4, 6, 10, 11]);

    document.getElementById("ul").removeAttribute("tabindex");

    assert_equals(content.childrenCount, 17,
                          "Making list unfocusable should make it hidden again");
    assert_true(content.isIgnored);
    testChildren(content, [0, 2, 4, 6, 11]);
});
</script>

</body>
</html>
